pyproject.toml
[project]
name = "academy-functions-front"
version = "0.1.0"
requires-python = "==3.13.5"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.0",
    "uvicorn>=0.38.0",
]
